+++ main.py
@@ -1,5 +1,5 @@
-from dl_bot.bot import DlBot
+from dl_bot.application import build_and_run
+
 
 if __name__ == "__main__":
-    bot = DlBot()
-    bot.build_and_run()
+    build_and_run()
